# apps/api/src/modules/communications/routes.ts
import type { FastifyInstance } from 'fastify';
import type { Queue } from 'bullmq';
import { z } from 'zod';
import { database } from '@zerochack/database';
import type { Environment } from '@zerochack/config';
import { communicationEventTypes } from '@zerochack/notifications';
import { renderEmailTemplate, validateEmailTemplate, type EmailProvider } from '@zerochack/email';
import { authenticate, requireOwnerMfa, requirePermission } from '../auth/security.js';
import { writeAudit } from '../auth/service.js';
import { ApiError } from '../../errors.js';

function parse<T>(schema:z.ZodType<T>,value:unknown):T{const result=schema.safeParse(value);if(!result.success)throw new ApiError(400,'VALIDATION_ERROR','Request validation failed',result.error.flatten());return result.data;}
const eventType=z.enum(communicationEventTypes);const uuid=z.string().uuid();

export async function communicationRoutes(app:FastifyInstance,options:{environment:Environment;email:EmailProvider;notificationsQueue?:Queue}):Promise<void>{
  app.addHook('preHandler',async(request)=>authenticate(request,options.environment));
  app.get('/notifications',async(request)=>{requirePermission(request,'profile.read');return database.notification.findMany({where:{tenantId:request.tenantId!,OR:[{recipientUserId:request.userId!},{recipientUserId:null}]},orderBy:{createdAt:'desc'},take:200});});
  app.patch('/notifications/:notificationId/read',async(request)=>{requirePermission(request,'profile.read');const id=parse(uuid,(request.params as {notificationId?:string}).notificationId);const result=await database.notification.updateMany({where:{id,tenantId:request.tenantId!,OR:[{recipientUserId:request.userId!},{recipientUserId:null}]},data:{readAt:new Date()}});if(!result.count)throw new ApiError(404,'NOTIFICATION_NOT_FOUND','Notification was not found');return{read:true};});
  app.get('/notification-preferences',async(request)=>{requirePermission(request,'notifications.manage');const saved=await database.notificationPreference.findMany({where:{tenantId:request.tenantId!,userId:request.userId!}});const map=new Map(saved.map((item)=>[item.eventType,item]));return{preferences:communicationEventTypes.map((type)=>map.get(type)??{eventType:type,inAppEnabled:true,emailEnabled:true})};});
  app.put('/notification-preferences/:eventType',async(request)=>{requirePermission(request,'notifications.manage');const type=parse(eventType,(request.params as {eventType?:string}).eventType);const input=parse(z.object({inAppEnabled:z.boolean(),emailEnabled:z.boolean()}),request.body);const preference=await database.notificationPreference.upsert({where:{tenantId_userId_eventType:{tenantId:request.tenantId!,userId:request.userId!,eventType:type}},create:{tenantId:request.tenantId!,userId:request.userId!,eventType:type,...input},update:input});await writeAudit({tenantId:request.tenantId,actorUserId:request.userId,requestId:request.id,action:'notification.preference_changed',resourceType:'notification_preference',resourceId:preference.id,ipAddress:request.ip,metadata:{eventType:type,...input}});return preference;});
  app.get('/owner/email/templates',async(request)=>{requireOwnerMfa(request);requirePermission(request,'email.templates.manage');return{templates:await database.emailTemplateVersion.findMany({orderBy:[{eventType:'asc'},{version:'desc'}]})};});
  app.post('/owner/email/templates/:eventType/versions',async(request,reply)=>{requireOwnerMfa(request);requirePermission(request,'email.templates.manage');const type=parse(eventType,(request.params as {eventType?:string}).eventType);const input=parse(z.object({subjectTemplate:z.string().min(1).max(240),bodyTemplate:z.string().min(1).max(20_000),enabled:z.boolean().default(true)}),request.body);try{validateEmailTemplate(input.subjectTemplate,240);validateEmailTemplate(input.bodyTemplate,20_000);}catch{throw new ApiError(400,'EMAIL_TEMPLATE_INVALID','Templates support only non-executable {{placeholder}} substitutions');}const created=await database.$transaction(async(tx)=>{const latest=await tx.emailTemplateVersion.aggregate({where:{eventType:type},_max:{version:true}});if(input.enabled)await tx.emailTemplateVersion.updateMany({where:{eventType:type,enabled:true},data:{enabled:false}});return tx.emailTemplateVersion.create({data:{eventType:type,version:(latest._max.version??0)+1,subjectTemplate:input.subjectTemplate,bodyTemplate:input.bodyTemplate,enabled:input.enabled,createdByUserId:request.userId!}});},{isolationLevel:'Serializable'});await writeAudit({tenantId:request.tenantId,actorUserId:request.userId,requestId:request.id,action:'email.template_version_created',resourceType:'email_template_version',resourceId:created.id,ipAddress:request.ip,metadata:{eventType:type,version:created.version,enabled:created.enabled}});return reply.code(201).send(created);});
  app.patch('/owner/email/templates/:versionId',async(request)=>{requireOwnerMfa(request);requirePermission(request,'email.templates.manage');const id=parse(uuid,(request.params as {versionId?:string}).versionId);const {enabled}=parse(z.object({enabled:z.boolean()}),request.body);const current=await database.emailTemplateVersion.findUnique({where:{id}});if(!current)throw new ApiError(404,'EMAIL_TEMPLATE_NOT_FOUND','Email template version was not found');await database.$transaction(async(tx)=>{if(enabled)await tx.emailTemplateVersion.updateMany({where:{eventType:current.eventType,enabled:true},data:{enabled:false}});await tx.emailTemplateVersion.update({where:{id},data:{enabled}});});await writeAudit({tenantId:request.tenantId,actorUserId:request.userId,requestId:request.id,action:'email.template_state_changed',resourceType:'email_template_version',resourceId:id,ipAddress:request.ip,metadata:{enabled}});return database.emailTemplateVersion.findUniqueOrThrow({where:{id}});});
  app.post('/owner/email/templates/preview',async(request)=>{requireOwnerMfa(request);requirePermission(request,'email.templates.manage');const input=parse(z.object({subjectTemplate:z.string().min(1).max(240),bodyTemplate:z.string().min(1).max(20_000),data:z.record(z.string(),z.unknown())}),request.body);try{return{subject:renderEmailTemplate(input.subjectTemplate,input.data,240),body:renderEmailTemplate(input.bodyTemplate,input.data)}}catch{throw new ApiError(400,'EMAIL_TEMPLATE_RENDER_FAILED','Template placeholders must resolve to scalar preview values');}});
  app.post('/owner/email/templates/:versionId/test-send',async(request)=>{requireOwnerMfa(request);requirePermission(request,'email.templates.manage');const id=parse(uuid,(request.params as {versionId?:string}).versionId);const input=parse(z.object({recipientEmail:z.email(),data:z.record(z.string(),z.unknown())}),request.body);const template=await database.emailTemplateVersion.findUnique({where:{id}});if(!template)throw new ApiError(404,'EMAIL_TEMPLATE_NOT_FOUND','Email template version was not found');let rendered;try{rendered={subject:renderEmailTemplate(template.subjectTemplate,input.data,240),text:renderEmailTemplate(template.bodyTemplate,input.data)}}catch{throw new ApiError(400,'EMAIL_TEMPLATE_RENDER_FAILED','Template placeholders must resolve to scalar test values');}const result=await options.email.send({to:input.recipientEmail,...rendered,idempotencyKey:`test-${request.id}`});await writeAudit({tenantId:request.tenantId,actorUserId:request.userId,requestId:request.id,action:'email.template_test_sent',resourceType:'email_template_version',resourceId:id,ipAddress:request.ip,metadata:{providerMessageId:result.providerMessageId}});return{sent:true,providerMessageId:result.providerMessageId};});
  app.get('/owner/email/deliveries',async(request)=>{requireOwnerMfa(request);requirePermission(request,'email.templates.manage');return{deliveries:await database.emailDelivery.findMany({select:{id:true,tenantId:true,recipientEmail:true,status:true,attempts:true,providerMessageId:true,lastErrorCode:true,nextAttemptAt:true,createdAt:true,deliveredAt:true,event:{select:{eventType:true}},templateVersion:{select:{version:true}}},orderBy:{createdAt:'desc'},take:500})};});
}